requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
mypy
pytest
pytest-django
python-dotenv
ruff
